#137 natural number trait class heierarchy completed.
