+++ Util/src/pParse/NatBase32Token.scala
@@ -2,8 +2,8 @@
 package ostrat
 package pParse
 
-/** Token for base 32 or duatricesimal natural number. All DecimalTokens are also legal base32 tokens, as well as all raw hexadecimal tokens. Base 32
- * , '0' .. '9', followed by 'a' .. 'v' with letter 'o' replaced with 'w'. Some Alpha numeric terms are valid TrigDualInts. */
+/** Token for base 32 natural number. All DecimalTokens are also legal base32 tokens, as well as all raw hexadecimal tokens. Base32 , '0' .. '9',
+ *  followed by 'a' .. 'v' with letter 'o' replaced with 'w'. Some Alpha numeric terms are valid TrigDualInts. */
 trait NatBase32Token extends NatToken
 {
   /** Uses the digitStr to calculate the natural number from the digitStr. */
