+++ Util/src/pParse/TokensOther.scala
@@ -1,0 +1,15 @@
+/* Copyright 2018-21 Richard Oliver. Licensed under Apache Licence version 2.0. */
+package ostrat
+package pParse
+
+/** The purpose of this token is for use at the beginning of a file, to make the the rest of the Statements, sub-statements. As if they were the
+ *  statements inside parenthesis. */
+case class HashAlphaToken(startPosn: TextPosn, srcStr: String) extends ExprToken
+{ override def subTypeStr: String = "HashAlpha"
+}
+
+case class UnderscoreToken(startPosn: TextPosn) extends EmptyExprToken with StatementMember
+{ def srcStr = "_"
+  override def exprName: String = "EmptyClauseExpr"
+  override def subTypeStr: String = "Underscore"
+}
